+++ packages/core/src/common/time.c
@@ -1,5 +1,6 @@
 #include <combo.h>
 
+#if defined(GAME_MM)
 u32 Time_Game2Linear(u8 day, u16 time)
 {
     u32 linear;
@@ -23,9 +24,35 @@ void Time_Linear2Game(u8* day, u16* time, u32 linear)
     *time = t;
 }
 
+u32 Time_FromHalfDay(u8 halfDay)
+{
+    return 0x10000 + (u32)halfDay * 0x8000;
+}
+
 u32 Time_LinearMoonCrash(void)
 {
-    return 0x10000 + 0x8000 * gSharedCustomSave.mm.halfDays;
+    int max;
+
+    /* Last of a BSR builtin makes it annoying */
+    max = 0;
+    for (int i = 0; i < 6; ++i)
+    {
+        if (gSharedCustomSave.mm.halfDays & (1 << i))
+            max = i + 1;
+    }
+
+    return Time_FromHalfDay(max);
+}
+
+u32 Time_LinearNewCycle(void)
+{
+    for (int i = 0; i < 6; ++i)
+    {
+        if (gSharedCustomSave.mm.halfDays & (1 << i))
+            return Time_FromHalfDay(i);
+    }
+
+    return Time_FromHalfDay(6);
 }
 
 int Time_IsMoonCrashLinear(u32 time)
@@ -41,3 +68,4 @@ int Time_IsMoonCrash(u8 day, u16 time)
 {
     return Time_IsMoonCrashLinear(Time_Game2Linear(day, time));
 }
+#endif